requirements.txt
requests
paho-mqtt
jsonpickle
pycryptodome